test(merge): add comprehensive edge case tests for cell merging and fix capacity overflow
- Added  to testXLMergeCells.cpp.
- Discovered and fixed a critical integer overflow bug in : Attempting to merge the entire worksheet () caused  to overflow a 32-bit integer, resulting in , which threw an incorrect . Fixed by casting to .
- Added  to verify the user's ability to locate the primary value of a merged block using  and  indexing.
- Aligned the test resilience with other modern OOXML libraries like Excelize.

+++ OpenXLSX/sources/XLWorksheet_Formatting.cpp
@@ -20,7 +20,7 @@ XLDataValidations& XLWorksheet::dataValidations()
 
 XLCellRange XLWorksheet::mergeCells(XLCellRange const& rangeToMerge, bool emptyHiddenCells)
 {
-    if (rangeToMerge.numRows() * rangeToMerge.numColumns() < 2) {
+    if (static_cast<uint64_t>(rangeToMerge.numRows()) * rangeToMerge.numColumns() < 2) {
         using namespace std::literals::string_literals;
         throw XLInputError("XLWorksheet::"s + __func__ + ": rangeToMerge must comprise at least 2 cells"s);
     }
